AI-App/backend-healtho/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
transformers==4.40.1
Pillow==10.3.0
torch==2.3.0